requirements.txt
# ============================================================================
# Core Frameworks (Corrected Names)
# ============================================================================
llama-index
smolagents
langchain
langchain-community

# ============================================================================
# LLM Providers & API Clients
# ============================================================================
google-genai
openai

# ============================================================================
# Unified Embeddings & Vector Store (Jina Only)
# ============================================================================
numpy<2
sentence-transformers       
chromadb             
llama-index-vector-stores-chroma

# ============================================================================
# Document & Web Processing (Minimalist & Efficient)
# ============================================================================
docling              
llama-index-readers-docling
llama-index-readers-file 
llama-index-readers-web  
html2text 
pypdf               
pandas
openpyxl                    
duckduckgo-search  
ddgs 
youtube-transcript-api      
llama-index-readers-youtube-transcript
matplotlib
seaborn
scikit-learn
scipy
beautifulsoup4
pyyaml
tqdm

# ============================================================================
# MCP & Tools
# ============================================================================
mcp               

# ============================================================================
# UI & Observability
# ============================================================================
streamlit
watchdog
pillow
python-multipart
langfuse
weave
openinference-instrumentation-smolagents

# ============================================================================
# Infrastructure
# ============================================================================
transformers
accelerate
kernels
einops
sentencepiece
nest-asyncio                
requests
python-dotenv
pydantic

# ============================================================================
# Diffusion Models & Audio
# ============================================================================
diffusers
soundfile
